+++ Runtime/Observables/Primaries/ObservableBool.cs
@@ -1,6 +1,6 @@
 ﻿using UnityEngine.Events;
 
-namespace TripleA.Observables.Primaries
+namespace TripleA.Utils.Observables.Primaries
 {
 	[System.Serializable]
 	public class ObservableBool : Observable<bool>