+++ Runtime/Singletons/PersistentSingleton.cs
@@ -1,7 +1,6 @@
 ﻿using UnityEngine;
-using UnityEngine.Serialization;
 
-namespace TripleA.Singletons
+namespace TripleA.Utils.Singletons
 {
 	public class PersistentSingleton<T> : GenericSingleton<T> where T : MonoBehaviour
 	{